third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/filtered-div-ref-svg-feimage-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not use cross-origin feimage content via CSS</title>
  <link rel="help" href="https://github.com/WICG/html-in-canvas">
  <link rel="author" href="mailto:schenney@chromium.org">
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <style>
    div {
      width: 100px;
      height: 100px;
      background: green;
    }
    #child-same {
      filter: url('#filter-same');
    }
    #child-cross {
      filter: url('#filter-cross');
    }
  </style>
</head>
<body>
  <svg width="0" height="0">
    <filter id="filter-same">
      <feimage href="https://{{location[host]}}/wpt_internal/html/canvas/drawElementImage/resources/red-100x100.png" />
    </filter>
    <filter id="filter-cross">
      <feImage href="https://{{hosts[alt][www]}}:{{ports[h2][0]}}/wpt_internal/html/canvas/drawElementImage/resources/red-100x100.png"/>
    </filter>
  </svg>

  <canvas id=canvas width="100" height="200" layoutsubtree>
    <div id="child-same"></div>
    <div id="child-cross"></div>
  </canvas>

  <script>
    window.onload = () => {
      test(function(t) {
        var context = canvas.getContext("2d", { willReadFrequently: true });
        context.drawElementImage(document.getElementById("child-same"), 0, 0);
        context.drawElementImage(document.getElementById("child-cross"), 0, 100);

        pixel = context.getImageData(50, 50, 1, 1).data;
        assert_equals(pixel[0], 255, "Same origin should draw");
        assert_equals(pixel[1], 0, "Same origin should draw");
        assert_equals(pixel[2], 0, "Same origin should draw");

        // Ignored filter is not applied at all
        pixel = context.getImageData(50, 150, 1, 1).data;
        assert_equals(pixel[0], 0, "Cross origin should not draw");
        assert_equals(pixel[1], 128, "Cross origin should not draw");
        assert_equals(pixel[2], 0, "Cross origin should not draw");

        t.done();
      });
    }
  </script>
</body>
</html>
